requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
spotipy==2.19.0
python-dotenv==0.19.0
Flask-SocketIO==5.1.1
gunicorn==20.1.0
gevent==22.10.2
gevent-websocket==0.10.1
eventlet==0.33.3
bidict==0.22.1
python-engineio==4.3.4
python-socketio==5.7.2
